done api crud student
done api crud student

+++ src/main/java/com/example/quanlysv/servlet/service/impl/StudentServiceImpl.java
@@ -1,15 +1,19 @@
 package com.example.quanlysv.servlet.service.impl;
 
+import com.example.quanlysv.servlet.common.Constant;
 import com.example.quanlysv.servlet.dao.IStudentDao;
 import com.example.quanlysv.servlet.dao.impl.StudentDaoImpl;
 import com.example.quanlysv.servlet.dto.request.BaseRequest;
+import com.example.quanlysv.servlet.dto.request.student.CreateOrEditStudentDTO;
 import com.example.quanlysv.servlet.dto.request.student.StudentDTO;
+import com.example.quanlysv.servlet.dto.request.student.StudentFilter;
 import com.example.quanlysv.servlet.dto.response.BaseResponse;
 import com.example.quanlysv.servlet.entity.StudentEntity;
 import com.example.quanlysv.servlet.service.IStudentService;
 import com.example.quanlysv.servlet.util.Convert;
 import org.apache.log4j.Logger;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 public class StudentServiceImpl implements IStudentService {
@@ -21,16 +25,18 @@ public class StudentServiceImpl implements IStudentService {
         studentDao = new StudentDaoImpl();
     }
 
+    private final Date date = new Date();
+
     @Override
-    public BaseResponse<?> findStudent(BaseRequest request) {
+    public BaseResponse<?> findStudent(StudentFilter request) {
         List<StudentDTO> result = new ArrayList<>();
         try{
 
-            if (request.getSortOrder().isEmpty()) {
-                request.setSortOrder("asc");
+            if (request.getBaseRequest().getSortOrder().isEmpty()) {
+                request.getBaseRequest().setSortOrder("asc");
             }
-            if(request.getSortField().isEmpty()){
-                request.setSortField("id_sv");
+            if(request.getBaseRequest().getSortField().isEmpty()){
+                request.getBaseRequest().setSortField("id_sv");
             }
 
             List<StudentEntity> list = studentDao.findStudent(request);
@@ -47,27 +53,91 @@ public class StudentServiceImpl implements IStudentService {
     }
 
     @Override
-    public void createOrUpdateStudent(StudentDTO studentDTO) {
+    public BaseResponse<?> createOrUpdateStudent(CreateOrEditStudentDTO studentDTO) {
         try {
+            // status =0 ==> create; status =1 ==> update (tao flag)
+            if(studentDTO.getStatus() == 0){
+                if(studentDao.existedByEmailAndIdSvAndPhone(studentDTO.getIdSv(),
+                        studentDTO.getEmailSv(), studentDTO.getPhoneSv())){
+                    return new BaseResponse.Builder<>()
+                            .setStatus(Constant.httpStatusErrorServer)
+                            .setMessage(Constant.messageStudentExist)
+                            .build();
+                };
+            }
+
             StudentEntity studentEntity = Convert.convertDTOToEntity(studentDTO, StudentEntity.class);
             if(studentEntity != null){
+                studentEntity.setNgayTao(date.getTime());
+                studentEntity.setNgaySua(date.getTime());
                 studentDao.createOrUpdateStudent(studentEntity);
+                return new BaseResponse.Builder<>()
+                        .setStatus(Constant.httpStatusOk)
+                        .setMessage(Constant.messageSuccess).build();
             }
+            return new BaseResponse.Builder<>()
+                    .setStatus(Constant.httpStatusErrorServer)
+                    .setMessage(Constant.messageFailed).build();
         } catch (IllegalAccessException e) {
-            throw new RuntimeException(e.getMessage());
+            log.error("failed by =>", e.getCause());
+            return new BaseResponse.Builder<>()
+                    .setStatus(Constant.httpStatusErrorServer)
+                    .setMessage(Constant.messageFailed).build();
         } catch (InstantiationException e) {
-            throw new RuntimeException(e.getMessage());
+            log.error("failed by =>", e.getCause());
+            return new BaseResponse.Builder<>()
+                    .setStatus(Constant.httpStatusErrorServer)
+                    .setMessage(Constant.messageFailed).build();
         }
     }
 
     @Override
-    public void deleteStudentById(String id) {
+    public BaseResponse<?> getStudentById(String id) {
+        try {
+            StudentEntity studentEntity = studentDao.getById(id);
+            if(studentEntity == null){
+                return new BaseResponse.Builder<>()
+                        .setStatus(Constant.httpStatusErrorServer)
+                        .setMessage(Constant.messageStudentNotFound).build();
+            }
+            StudentDTO studentDTO = Convert.convertEntityToDTO(studentEntity, StudentDTO.class);
+            return new BaseResponse.Builder<>()
+                    .setStatus(Constant.httpStatusOk)
+                    .setData(studentDTO)
+                    .setMessage(Constant.messageSuccess).build();
+
+        }catch (Exception e){
+            log.error(e.getMessage());
+            return new BaseResponse.Builder<>()
+                    .setStatus(Constant.httpStatusErrorServer)
+                    .setMessage(Constant.messageStudentNotFound).build();
+        }
+
+    }
+
+    @Override
+    public BaseResponse<?> deleteStudentById(String id) {
         try{
+            StudentEntity studentEntity = studentDao.getById(id);
+            if(studentEntity == null){
+                return new BaseResponse.Builder<>()
+                        .setStatus(Constant.httpStatusErrorServer)
+                        .setMessage(Constant.messageStudentNotFound).build();
+            }
             if(!id.trim().isEmpty()) {
                 studentDao.deleteStudentById(id.trim());
+                return new BaseResponse.Builder<>()
+                        .setStatus(Constant.httpStatusOk)
+                        .setMessage(Constant.messageSuccess).build();
             }
+            return new BaseResponse.Builder<>()
+                    .setStatus(Constant.httpStatusErrorServer)
+                    .setMessage(Constant.messageStudentNotFound).build();
         }catch (Exception e){
-            throw new RuntimeException("create student failed: "+ e.getMessage());
+            log.error(e.getMessage());
+            return new BaseResponse.Builder<>()
+                    .setStatus(Constant.httpStatusErrorServer)
+                    .setMessage(Constant.messageStudentNotFound).build();
         }
     }
 }
